feat(EMS-2120): move insurance eligibility session creation to the new first page
