+++ src/ui/server/controllers/insurance/eligibility/buyer-country/index.test.ts
@@ -88,42 +88,6 @@ describe('controllers/insurance/eligibility/buyer-country', () => {
       expect(res.render).toHaveBeenCalledWith(TEMPLATE, expectedVariables);
     });
 
-    describe('when a there is no submittedData in req.session', () => {
-      it('should add empty submittedData.insuranceEligibility to the session', async () => {
-        // @ts-ignore
-        req.session = {};
-
-        await get(req, res);
-
-        const expected = {
-          ...req.session,
-          submittedData: {
-            insuranceEligibility: {},
-          },
-        };
-
-        expect(req.session).toEqual(expected);
-      });
-    });
-
-    describe('when a there is no insuranceEligibility in req.session.submittedData', () => {
-      it('should add empty submittedData.insuranceEligibility to the session and retain existing req.session.submittedData', async () => {
-        // @ts-ignore
-        req.session.submittedData = {
-          quoteEligibility: {},
-        };
-
-        await get(req, res);
-
-        const expected = {
-          ...req.session.submittedData,
-          insuranceEligibility: {},
-        };
-
-        expect(req.session.submittedData).toEqual(expected);
-      });
-    });
-
     describe('when a country has been submitted', () => {
       it('should render template with countries mapped to submitted country', async () => {
         req.session.submittedData = mockSession.submittedData;
